Move 3rd party integrations into dist folder
