+++ FURTHER-INTEGRATIONS.md
@@ -6,13 +6,15 @@
 
 ```js
 // mixin.core.js
-const { Mixin } = require('hops');
-const BabelTimingPlugin = require('babel-timing/webpack/plugin');
+const {Mixin} = require('hops');
+const BabelTimingPlugin = require('babel-timing/dist/webpack/plugin');
 
 class BabelTimingMixin extends Mixin {
   configureBuild(webpackConfig, loaderConfigs, target) {
-    if(target === 'build') {
-      loaderConfigs.jsLoaderConfig.options.customize = require.resolve('babel-timing/webpack/babel-loader-customize');
+    if (target === 'build') {
+      loaderConfigs.jsLoaderConfig.options.customize = require.resolve(
+        'babel-timing/dist/webpack/babel-loader-customize'
+      );
       webpackConfig.plugins.push(new BabelTimingPlugin());
     }
   }
