+++ src/jest/__tests__/__fixtures__/.babelrc
