+++ src/webpack/__tests__/__fixtures__/.babelrc
