+++ src/webpack/__tests__/__fixtures__/entry.js
