+++ src/webpack/__tests__/__fixtures__/sum.js
