+++ src/webpack/__tests__/__fixtures__/webpack.config.js
